requirements.txt
openai==1.12.0
PyYAML==6.0.1